Set args in TestCase.assertParseArgsRaises calls.
If the `args` parameter is not set when calling TestCase.assertParseArgsRaises, `args` is set to `sys.argv`, which makes the test output dependent on how the test suit was called. For example, it would succeed for `pytest`, but fail with `pytest -v` or `pytest -k testConfigOrEnvValueErrors`.

Change signature of assertParseArgsRaises() to ensure args is always present, and all other parameters are named.