Allow setting counters in configfile.
E.g. `action=count`, called `verbose`/`v`

You can still set `-v -v -v` (or `-vvv` on the command line).
And you can now also set:
    verbose=3
in the configuration file.

Added two tests to verify current behaviour.
Note that it still is not possible to specify `-v=3` on the command line (or three `verbose` lines in the configuration file).

Credits to @christensonb for suggesting this fix in #138.
Closes #138
Closes #175